+++ R/tm_t_events.R
@@ -8,8 +8,9 @@
 #' @param event_type (`character`)\cr type of event that is summarized (e.g. adverse event, treatment).
 #'   Default is "event".
 #' @param sort_criteria (`character`)\cr how to sort the final table. Default option `freq_desc` sorts
-#'   by decreasing total number of patients with event. Alternative option `alpha` sorts events
+#'   on column `sort_freq_col` by decreasing number of patients with event. Alternative option `alpha` sorts events
 #'   alphabetically.
+#' @param sort_freq_col (`character`)\cr column to sort by frequency on if `sort_criteria` is set to `freq_desc`.
 #'
 #' @seealso [tm_t_events()]
 #' @keywords internal
@@ -25,6 +26,7 @@ template_events <- function(dataname,
                             total_label = "All Patients",
                             event_type = "event",
                             sort_criteria = c("freq_desc", "alpha"),
+                            sort_freq_col = total_label,
                             prune_freq = 0,
                             prune_diff = 0,
                             drop_arm_levels = TRUE,
@@ -362,10 +364,17 @@ template_events <- function(dataname,
     }
   } else {
     # Sort by decreasing frequency.
+    sort_list <- add_expr(
+      sort_list,
+      substitute(
+        expr = idx_split_col <- which(sapply(col_paths(result), tail, 1) == sort_freq_col),
+        env = list(sort_freq_col = sort_freq_col)
+      )
+    )
 
     # When the "All Patients" column is present we only use that for scoring.
     scorefun_hlt <- if (add_total) {
-      quote(cont_n_onecol(ncol(result)))
+      quote(cont_n_onecol(idx_split_col))
     } else {
       quote(cont_n_allcols)
     }
@@ -491,6 +500,7 @@ tm_t_events <- function(label,
                         total_label = "All Patients",
                         event_type = "event",
                         sort_criteria = c("freq_desc", "alpha"),
+                        sort_freq_col = total_label,
                         prune_freq = 0,
                         prune_diff = 0,
                         drop_arm_levels = TRUE,
@@ -533,6 +543,7 @@ tm_t_events <- function(label,
         event_type = event_type,
         label = label,
         total_label = total_label,
+        sort_freq_col = sort_freq_col,
         basic_table_args = basic_table_args
       )
     ),
@@ -592,7 +603,7 @@ ui_t_events_byterm <- function(id, ...) {
           selected = a$sort_criteria,
           multiple = FALSE
         ),
-        shiny::helpText("Pruning Options"),
+        shiny::helpText(strong("Pruning Options:")),
         shiny::numericInput(
           inputId = ns("prune_freq"),
           label = "Minimum Incidence Rate(%) in any of the treatment groups",
@@ -636,6 +647,7 @@ srv_t_events_byterm <- function(id,
                                 drop_arm_levels,
                                 label,
                                 total_label,
+                                sort_freq_col,
                                 basic_table_args) {
   with_reporter <- !missing(reporter) && inherits(reporter, "Reporter")
   with_filter <- !missing(filter_panel_api) && inherits(filter_panel_api, "FilterPanelAPI")
@@ -757,6 +769,7 @@ srv_t_events_byterm <- function(id,
         total_label = total_label,
         event_type = event_type,
         sort_criteria = input$sort_criteria,
+        sort_freq_col = sort_freq_col,
         prune_freq = input$prune_freq / 100,
         prune_diff = input$prune_diff / 100,
         drop_arm_levels = input$drop_arm_levels,